+++ gossip/nostr-gossip-memory/CHANGELOG.md
@@ -23,6 +23,12 @@
 
 -->
 
+## Unreleased
+
+### Changed
+
+- Replace `Mutex` with `RwLock` for better concurrency (https://github.com/rust-nostr/nostr/pull/1126)
+
 ## v0.44.0 - 2025/11/06
 
 First release.